backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
google-generativeai==0.8.5
python-dotenv==1.0.1
pillow==11.3.0
